Socket implementation (WIP)
Connected video feed and audio feed using socket library. As of now, the audio recording python script sends whether or not the recording of the first second has an rms value over 1000 to the video script. A success is an rms value over 1000, where the video script responds by opening an unedited opencv feed, and a failure is when the rms value is below 1000, where the video script responds by opening an opencv feed in grayscale. The goal is to always have the video feed playing, and actively responding to loud noises by switching between grayscale and unedited feeds.